requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
Flask==2.0.3
frozenlist==1.3.0
future==0.18.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
line-bot-sdk==2.2.1
MarkupSafe==2.0.1
multidict==6.0.2
pip==21.2.2
requests==2.27.1
setuptools==61.2.0
urllib3==1.26.9
Werkzeug==2.0.3
wheel==0.37.1
wincertstore==0.2
yarl==1.7.2

redis=4.2.2
pydantic=1.9.0